requirements.txt
annotated-types==0.6.0 # Pydantic
pydantic==2.6.4 # Pydantic
pydantic-core==2.16.3 # Pydantic
typing-extensions==4.10.0 # Pydantic

anyio==4.3.0 # FastAPI
fastapi==0.110.1 # FastAPI
idna==3.6 # FastAPI
sniffio==1.3.1 # FastAPI
starlette==0.37.2 # FastAPI

fastapi-pagination==0.12.21 # Pagination
sqlalchemy==2.0.29 # SQLAlchemy

python-dotenv==1.0.1 # Environment variables
